Build a stub libobs instead of linking real one
By building a fake stubbed out libobs ourselves, it's much easier to
build this. The actual libobs is not required anymore.

Additionally we now do CI builds for macOS aarch64.

Also this now references the master branch of livesplit-core rather than
some in-progress branch that doesn't exist anymore.
