+++ .gitignore
@@ -1,1 +1,4 @@
 /target
+/libobs.so
+/libobs.dylib
+/obs.lib
