+++ Cargo.toml
@@ -6,11 +6,14 @@ edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[workspace]
+members = ["obs"]
+
 [lib]
 crate-type = ["cdylib"]
 
 [dependencies]
-livesplit-core = { git = "https://github.com/CryZe/livesplit-core", branch = "path-renderer", features = ["software-rendering", "font-loading"] }
+livesplit-core = { git = "https://github.com/LiveSplit/livesplit-core", features = ["software-rendering", "font-loading"] }
 once_cell = "1.5.2"
 
 [profile.dev.package."*"]
