+++ README.md
@@ -23,11 +23,3 @@ Download the latest release for your operating system from the
 - Right click your `OBS` -> Options -> Show in Finder
 - Right click the `OBS.app` -> Show Package Contents
 - Drag `obs-livesplit-one.so` into `Contents/PlugIns`
-
-## Developer Notes
-
-If you add a new function to `ffi.rs`, also add it to `exports.def` and run
-(needs to be done in the Visual Studio console):
-```sh
-lib /def:exports.def /OUT:obs.lib /MACHINE:x64
-```
