--- exports.def
@@ -1,27 +1,0 @@
-LIBRARY obs
-EXPORTS
-obs_register_source_s
-gs_texture_create
-obs_source_draw
-obs_enter_graphics
-obs_leave_graphics
-gs_texture_set_image
-obs_hotkey_register_source
-obs_properties_create
-obs_properties_add_path
-obs_data_get_string
-blog
-obs_properties_add_int
-obs_data_get_int
-gs_texture_destroy
-gs_draw_sprite
-gs_effect_get_param_by_name
-gs_effect_get_technique
-gs_effect_set_texture
-gs_technique_begin
-gs_technique_begin_pass
-gs_technique_end
-gs_technique_end_pass
-obs_get_base_effect
-obs_data_set_default_int
-obs_properties_add_button
